refactor anim component code and update binder logic to use paths instead of nodes
