+++ src/anim/controller/anim-controller.js
@@ -1,488 +1,13 @@
-import { AnimClip } from '../../../anim/anim-clip.js';
+import { AnimClip } from '../evaluator/anim-clip.js';
+import { AnimState } from './anim-state.js';
+import { AnimBlendTree } from './anim-blend-tree.js';
+import { AnimTransition } from './anim-transition.js';
 
 import {
-    ANIM_GREATER_THAN, ANIM_LESS_THAN, ANIM_GREATER_THAN_EQUAL_TO, ANIM_LESS_THAN_EQUAL_TO, ANIM_EQUAL_TO, ANIM_NOT_EQUAL_TO,
     ANIM_INTERRUPTION_NONE, ANIM_INTERRUPTION_PREV, ANIM_INTERRUPTION_NEXT, ANIM_INTERRUPTION_PREV_NEXT, ANIM_INTERRUPTION_NEXT_PREV,
     ANIM_PARAMETER_TRIGGER,
-    ANIM_STATE_START, ANIM_STATE_END, ANIM_STATE_ANY,
-    ANIM_BLEND_1D, ANIM_BLEND_2D_DIRECTIONAL, ANIM_BLEND_2D_CARTESIAN, ANIM_BLEND_DIRECT
+    ANIM_STATE_START, ANIM_STATE_END, ANIM_STATE_ANY
 } from './constants.js';
-
-/**
- * @private
- * @class
- * @name pc.AnimNode
- * @classdesc AnimNodes are used to represent a single animation track in the current state. Each state can contain multiple AnimNodes, in which case they are stored in a BlendTree hierarchy, which will control the weight (contribution to the states final animation) of it's child AnimNodes.
- * @description Create a new AnimNode.
- * @param {pc.AnimState} state - The AnimState that this BlendTree belongs to.
- * @param {pc.BlendTree|null} parent - The parent of the AnimNode. If not null, the AnimNode is stored as part of a pc.BlendTree hierarchy.
- * @param {string} name - The name of the AnimNode. Used when assigning a pc.AnimTrack to it.
- * @param {number[]|pc.Vec2} point - The coordinate/vector thats used to determine the weight of this node when it's part of a pc.BlendTree.
- * @param {number} [speed] - The speed that its pc.AnimTrack should play at. Defaults to 1.
- */
-class AnimNode {
-    constructor(state, parent, name, point, speed = 1) {
-        this._state = state;
-        this._parent = parent;
-        this._name = name;
-        this._point = Array.isArray(point) ? new pc.Vec2(point) : point;
-        this._speed = speed;
-        this._weight = 1.0;
-        this._animTrack = null;
-    }
-
-    get parent() {
-        return this._parent;
-    }
-
-    get name() {
-        return this._name;
-    }
-
-    get path() {
-        return this._parent ? this._parent.path + '.' + this._name : this._name;
-    }
-
-    get point() {
-        return this._point;
-    }
-
-    get weight() {
-        return this._parent ? this._parent.weight * this._weight : this._weight;
-    }
-
-    set weight(value) {
-        this._weight = value;
-    }
-
-    get normalizedWeight() {
-        var totalWeight = this._state.totalWeight;
-        if (totalWeight === 0.0) return 0.0;
-        return this.weight / totalWeight;
-    }
-
-    get speed() {
-        return this._speed;
-    }
-
-    get animTrack() {
-        return this._animTrack;
-    }
-
-    set animTrack(value) {
-        this._animTrack = value;
-    }
-}
-
-// Maths helper functions
-function between(num, a, b, inclusive) {
-    var min = Math.min(a, b),
-        max = Math.max(a, b);
-    return inclusive ? num >= min && num <= max : num > min && num < max;
-}
-
-function getAngleRad(a, b) {
-    return Math.atan2(a.x * b.y - a.y * b.x, a.x * b.x + a.y * b.y);
-}
-
-function clamp(num, min, max) {
-    return num <= min ? min : num >= max ? max : num;
-}
-
-/**
- * @private
- * @class
- * @name pc.BlendTree
- * @classdesc BlendTrees are used to store and blend multiple AnimNodes together. BlendTrees can be the child of other BlendTrees, in order to create a hierarchy of AnimNodes. It takes a blend type as an argument which defines which function should be used to determine the weights of each of it's children, based on the current parameter value.
- * @description Create a new BlendTree.
- * @param {pc.AnimState} state - The AnimState that this BlendTree belongs to.
- * @param {pc.BlendTree|null} parent - The parent of the BlendTree. If not null, the AnimNode is stored as part of a pc.BlendTree hierarchy.
- * @param {string} name - The name of the BlendTree. Used when assigning a pc.AnimTrack to its children.
- * @param {number|pc.Vec2} point - The coordinate/vector thats used to determine the weight of this node when it's part of a pc.BlendTree.
- * @param {string} type - Determines which blending algorithm is used to calculate the weights of its child nodes. One of pc.ANIM_BLEND_*.
- * @param {string[]} parameters - The anim component parameters which are used to calculate the current weights of the blend trees children.
- * @param {object[]} children - The child nodes that this blend tree should create. Can either be of type pc.AnimNode or pc.BlendTree.
- * @param {Function} findParameter - Used at runtime to get the current parameter values.
- */
-class BlendTree extends AnimNode {
-    constructor(state, parent, name, point, type, parameters, children, findParameter) {
-        super(state, parent, name, point);
-        this._type = type;
-        this._parameters = parameters;
-        this._parameterValues = null;
-        this._children = [];
-        this._findParameter = findParameter;
-        for (var i = 0; i < children.length; i++) {
-            var child = children[i];
-            if (child.children) {
-                this._children.push(new BlendTree(state, this, child.name, child.point, child.type, child.parameter ? [child.parameter] : child.parameters, child.children, findParameter));
-            } else {
-                this._children.push(new AnimNode(state, this, child.name, child.point, child.speed));
-            }
-        }
-    }
-
-    get parent() {
-        return this._parent;
-    }
-
-    get name() {
-        return this._name;
-    }
-
-    get point() {
-        return this._point;
-    }
-
-    get weight() {
-        this.calculateWeights();
-        return this._parent ? this._parent.weight * this._weight : this._weight;
-    }
-
-    set weight(value) {
-        this._weight = value;
-    }
-
-    get speed() {
-        return this._speed;
-    }
-
-    getChild(name) {
-        for (var i = 0; i < this._children.length; i++) {
-            if (this._children[i].name === name) return this._children[i];
-        }
-    }
-
-    calculateWeights() {
-        var i, j, p, pi, pj, pip, pipj, parameterValues, minj, result, weightSum;
-        switch (this._type) {
-            case ANIM_BLEND_1D: {
-                var parameterValue = this._findParameter(this._parameters[0]).value;
-                if (this._parameterValues && (this._parameterValues[0] === parameterValue)) {
-                    return;
-                }
-                this._parameterValues = [parameterValue];
-                this._children[0].weight = 0.0;
-                for (i = 0; i < this._children.length - 1; i++) {
-                    var child1 = this._children[i];
-                    var child2 = this._children[i + 1];
-                    if (between(parameterValue, child1.point, child2.point, true)) {
-                        var child2Distance = Math.abs(child1.point - child2.point);
-                        var parameterDistance = Math.abs(child1.point - parameterValue);
-                        var weight = (child2Distance - parameterDistance) / child2Distance;
-                        child1.weight = weight;
-                        child2.weight = (1.0 - weight);
-                    } else {
-                        child2.weight = 0.0;
-                    }
-                }
-                break;
-            }
-            case ANIM_BLEND_2D_CARTESIAN: {
-                parameterValues = this._parameters.map(function (param) {
-                    return this._findParameter(param).value;
-                }.bind(this));
-                if (this._parameterValues && (this._parameterValues.equals(parameterValues))) {
-                    return;
-                }
-                this._parameterValues = parameterValues;
-                p = new pc.Vec2(this._parameterValues);
-
-                weightSum = 0.0;
-                for (i = 0; i < this._children.length; i++) {
-                    pi = this._children[i].point.clone();
-                    minj = Number.MAX_VALUE;
-                    for (j = 0; j < this._children.length; j++) {
-                        if (i === j) continue;
-                        pj = this._children[j].point.clone();
-                        pipj = pj.clone().sub(pi);
-                        pip = p.clone().sub(pi);
-                        result = clamp(1.0 - (pip.clone().dot(pipj) / pipj.lengthSq()), 0.0, 1.0);
-                        if (result < minj) minj = result;
-                    }
-                    this._children[i].weight = minj;
-                    weightSum += minj;
-                }
-                for (i = 0; i < this._children.length; i++) {
-                    this._children[i].weight = this._children[i]._weight / weightSum;
-                }
-                break;
-            }
-            case ANIM_BLEND_2D_DIRECTIONAL: {
-                parameterValues = this._parameters.map(function (param) {
-                    return this._findParameter(param).value;
-                }.bind(this));
-                if (this._parameterValues && (this._parameterValues.equals(parameterValues))) {
-                    return;
-                }
-                this._parameterValues = parameterValues;
-                p = new pc.Vec2(this._parameterValues);
-
-
-                weightSum = 0.0;
-                for (i = 0; i < this._children.length; i++) {
-                    pi = this._children[i].point.clone();
-                    minj = Number.MAX_VALUE;
-                    for (j = 0; j < this._children.length; j++) {
-                        if (i === j) continue;
-                        pj = this._children[j].point.clone();
-                        var pipAngle = getAngleRad(pi, p);
-                        var pipjAngle = getAngleRad(pi, pj);
-                        pipj = new pc.Vec2((pj.length() - pi.length()) / ((pj.length() + pi.length()) / 2), pipjAngle * 2.0);
-                        pip = new pc.Vec2((p.length() - pi.length()) / ((pj.length() + pi.length()) / 2), pipAngle * 2.0);
-                        result = clamp(1.0 - Math.abs((pip.clone().dot(pipj) / pipj.lengthSq())), 0.0, 1.0);
-                        if (result < minj) minj = result;
-                    }
-                    this._children[i].weight = minj;
-                    weightSum += minj;
-                }
-                for (i = 0; i < this._children.length; i++) {
-                    this._children[i].weight = this._children[i]._weight / weightSum;
-                }
-                break;
-            }
-            case ANIM_BLEND_DIRECT: {
-                parameterValues = this._parameters.map(function (param) {
-                    return this._findParameter(param).value;
-                }.bind(this));
-                if (this._parameterValues === parameterValues) {
-                    return;
-                }
-                this._parameterValues = parameterValues;
-                weightSum = 0.0;
-                for (i = 0; i < this._children.length; i++) {
-                    weightSum += clamp(this._parameterValues[i], 0.0, Number.MAX_VALUE);
-                }
-                for (i = 0; i < this._children.length; i++) {
-                    this._children[i].weight = clamp(this._parameterValues[i], 0.0, Number.MAX_VALUE) / weightSum;
-                }
-                break;
-            }
-        }
-    }
-
-    getNodeCount() {
-        var count = 0;
-        for (var i = 0; i < this._children.length; i++) {
-            var child = this._children[i];
-            if (child.constructor === BlendTree) {
-                count += this._children[i].getNodeCount();
-            } else {
-                count++;
-            }
-        }
-        return count;
-    }
-}
-
-/**
- * @private
- * @class
- * @name pc.AnimState
- * @classdesc Defines a single state that the controller can be in. Each state contains either a single AnimNode or a BlendTree of multiple AnimNodes, which will be used to animate the Entity while the state is active. An AnimState will stay active and play as long as there is no AnimTransition with it's conditions met that has that AnimState as it's source state.
- * @description Create a new AnimState.
- * @param {pc.AnimController} controller - The controller this AnimState is associated with.
- * @param {string} name - The name of the state. Used to find this state when the controller transitions between states and links animations.
- * @param {number} speed - The speed animations in the state should play at. Individual pc.AnimNodes can override this value.
- * @param {boolean} loop - Determines whether animations in this state should loop.
- * @param {object|null} blendTree - If supplied, the AnimState will recursively build a pc.BlendTree hierarchy, used to store, blend and play multiple animations.
- */
-class AnimState {
-    constructor(controller, name, speed, loop, blendTree) {
-        this._controller = controller;
-        this._name = name;
-        this._animations = {};
-        this._animationList = [];
-        this._speed = speed || 1.0;
-        this._loop = loop === undefined ? true : loop;
-        var findParameter = this._controller.findParameter.bind(this._controller);
-        if (blendTree) {
-            this._blendTree = new BlendTree(this, null, name, 1.0, blendTree.type, blendTree.parameter ? [blendTree.parameter] : blendTree.parameters, blendTree.children, findParameter);
-        } else {
-            this._blendTree = new AnimNode(this, null, name, 1.0, speed);
-        }
-    }
-
-    _getNodeFromPath(path) {
-        var currNode = this._blendTree;
-        for (var i = 1; i < path.length; i++) {
-            currNode = currNode.getChild(path[i]);
-        }
-        return currNode;
-    }
-
-    addAnimation(path, animTrack) {
-        var indexOfAnimation = this._animationList.findIndex(function (animation) {
-            return animation.path === path;
-        });
-        if (indexOfAnimation >= 0) {
-            this._animationList[indexOfAnimation].animTrack = animTrack;
-        } else {
-            var node = this._getNodeFromPath(path);
-            node.animTrack = animTrack;
-            this._animationList.push(node);
-        }
-    }
-
-    get name() {
-        return this._name;
-    }
-
-    get animations() {
-        return this._animationList;
-    }
-
-    set animations(value) {
-        this._animationList = value;
-    }
-
-    get speed() {
-        return this._speed;
-    }
-
-    get loop() {
-        return this._loop;
-    }
-
-    get nodeCount() {
-        if (!this._blendTree || !(this._blendTree.constructor === BlendTree)) return 1;
-        return this._blendTree.getNodeCount();
-    }
-
-    get playable() {
-        return (this.name === ANIM_STATE_START || this.name === ANIM_STATE_END || this.name === ANIM_STATE_ANY || this.animations.length === this.nodeCount);
-    }
-
-    get looping() {
-        if (this.animations.length > 0) {
-            var trackClipName = this.name + '.' + this.animations[0].animTrack.name;
-            var trackClip = this._controller.animEvaluator.findClip(trackClipName);
-            if (trackClip) {
-                return trackClip.loop;
-            }
-        }
-        return false;
-    }
-
-    get totalWeight() {
-        var sum = 0;
-        var i;
-        for (i = 0; i < this.animations.length; i++) {
-            sum += this.animations[i].weight;
-        }
-        return sum;
-    }
-
-    get timelineDuration() {
-        var duration = 0;
-        var i;
-        for (i = 0; i < this.animations.length; i++) {
-            var animation = this.animations[i];
-            if (animation.animTrack.duration > duration) {
-                duration = animation.animTrack.duration;
-            }
-        }
-        return duration;
-    }
-}
-
-/**
- * @private
- * @class
- * @name pc.AnimTransition
- * @classdesc AnimTransitions represent connections in the controllers state graph between AnimStates. During each frame, the controller tests to see if any of the AnimTransitions have the current AnimState as their source (from) state. If so and the AnimTransitions parameter based conditions are met, the controller will transition to the destination state.
- * @description Create a new AnimTransition.
- * @param {pc.AnimController} controller - The controller this AnimTransition is associated with.
- * @param {string} from - The state that this transition will exit from.
- * @param {string} to - The state that this transition will transition to.
- * @param {number} time - The duration of the transition in seconds.
- * @param {number} priority - Used to sort all matching transitions in ascending order. The first transition in the list will be selected.
- * @param {object[]} conditions - A list of conditions which must pass for this transition to be used.
- * @param {number} exitTime - If provided, this transition will only be active for the exact frame during which the source states progress passes the time specified. Given as a normalised value of the source states duration. Values less than 1 will be checked every animation loop.
- * @param {number} transitionOffset - If provided, the destination state will begin playing its animation at this time. Given in normalised time, based on the states duration & must be between 0 and 1.
- * @param {string} interruptionSource - Defines whether another transition can interrupt this one and which of the current or previous states transitions can do so. One of pc.ANIM_INTERRUPTION_*.
- */
-class AnimTransition {
-    constructor(controller, from, to, time, priority, conditions, exitTime, transitionOffset, interruptionSource = ANIM_INTERRUPTION_NONE) {
-        this._controller = controller;
-        this._from = from;
-        this._to = to;
-        this._time = time;
-        this._priority = priority;
-        this._conditions = conditions || [];
-        this._exitTime = exitTime || null;
-        this._transitionOffset = transitionOffset || null;
-        this._interruptionSource = interruptionSource;
-    }
-
-    get from() {
-        return this._from;
-    }
-
-    get to() {
-        return this._to;
-    }
-
-    get time() {
-        return this._time;
-    }
-
-    get priority() {
-        return this._priority;
-    }
-
-    get conditions() {
-        return this._conditions;
-    }
-
-    get exitTime() {
-        return this._exitTime;
-    }
-
-    get transitionOffset() {
-        return this._transitionOffset;
-    }
-
-    get interruptionSource() {
-        return this._interruptionSource;
-    }
-
-    get hasExitTime() {
-        return !!this.exitTime;
-    }
-
-    get hasConditionsMet() {
-        var conditionsMet = true;
-        var i;
-        for (i = 0; i < this.conditions.length; i++) {
-            var condition = this.conditions[i];
-            var parameter = this._controller.findParameter(condition.parameterName);
-            switch (condition.predicate) {
-                case ANIM_GREATER_THAN:
-                    conditionsMet = conditionsMet && parameter.value > condition.value;
-                    break;
-                case ANIM_LESS_THAN:
-                    conditionsMet = conditionsMet && parameter.value < condition.value;
-                    break;
-                case ANIM_GREATER_THAN_EQUAL_TO:
-                    conditionsMet = conditionsMet && parameter.value >= condition.value;
-                    break;
-                case ANIM_LESS_THAN_EQUAL_TO:
-                    conditionsMet = conditionsMet && parameter.value <= condition.value;
-                    break;
-                case ANIM_EQUAL_TO:
-                    conditionsMet = conditionsMet && parameter.value === condition.value;
-                    break;
-                case ANIM_NOT_EQUAL_TO:
-                    conditionsMet = conditionsMet && parameter.value !== condition.value;
-                    break;
-            }
-            if (!conditionsMet)
-                return conditionsMet;
-        }
-        return conditionsMet;
-    }
-}
 
 /**
  * @private
@@ -990,7 +515,7 @@ class AnimController {
                 }
             }
         } else {
-            if (this.activeState._blendTree.constructor === BlendTree) {
+            if (this.activeState._blendTree.constructor === AnimBlendTree) {
                 state = this.activeState;
                 for (i = 0; i < state.animations.length; i++) {
                     animation = state.animations[i];
