+++ src/anim/evaluator/anim-cache.js
@@ -1,6 +1,6 @@
-import { math } from '../math/math.js';
+import { math } from '../../math/math.js';
 
-import { INTERPOLATION_CUBIC, INTERPOLATION_LINEAR, INTERPOLATION_STEP } from './constants.js';
+import { INTERPOLATION_CUBIC, INTERPOLATION_LINEAR, INTERPOLATION_STEP } from '../constants.js';
 
 /**
  * @private
