+++ src/anim/evaluator/anim-clip.js
