+++ src/anim/evaluator/anim-curve.js
