+++ src/anim/evaluator/anim-snapshot.js
