+++ src/framework/components/anim/component.js
@@ -1,16 +1,16 @@
 import { Asset } from '../../../asset/asset.js';
 
-import { AnimEvaluator } from '../../../anim/anim-evaluator.js';
+import { AnimEvaluator } from '../../../anim/evaluator/anim-evaluator.js';
+import { AnimController } from '../../../anim/controller/anim-controller.js';
 
 import { Component } from '../component.js';
 
 import {
     ANIM_PARAMETER_BOOLEAN, ANIM_PARAMETER_FLOAT, ANIM_PARAMETER_INTEGER, ANIM_PARAMETER_TRIGGER, ANIM_STATE_START, ANIM_STATE_END, ANIM_STATE_ANY
-} from './constants.js';
-import { AnimComponentBinder } from './binder.js';
-import { AnimComponentLayer } from './layer.js';
-import { AnimController } from './controller.js';
-import { AnimStateGraph } from './state-graph.js';
+} from '../../../anim/controller/constants.js';
+import { AnimComponentBinder } from './component-binder.js';
+import { AnimComponentLayer } from './component-layer.js';
+import { AnimStateGraph } from '../../../resources/anim-state-graph.js';
 
 /**
  * @private
